pyproject.toml
[project]
name = "insanely-fast-whisper-rocm"
version = "2.1.5"
description = "ROCm-first Whisper transcription with FastAPI, CLI, and optional Gradio UI"
requires-python = ">=3.10,<3.11"
authors = [
    {name = "elvee"},
]
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "python-multipart==0.0.22",
    "transformers==4.57.3",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "click==8.3.1",
    "gradio==6.7.0",
    "pydub==0.25.1",
    "numpy==2.2.6",
    "accelerate==1.12.0",
    "ffmpeg-python==0.2.0",
    "stable-ts==2.19.1",
    "demucs==4.0.1",
    "huggingface-hub==0.36.0",
    "soundfile==0.13.1",
]


[project.scripts]
insanely-fast-whisper-cli = "insanely_fast_whisper_rocm.cli.cli:main"
insanely-fast-whisper-webui = "insanely_fast_whisper_rocm.webui.app:launch_webui"
insanely-fast-whisper-rocm = "insanely_fast_whisper_rocm.__main__:main"
insanely-fast-whisper-rocm-preflight = "insanely_fast_whisper_rocm.utils.rocm_preflight:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.metadata]
allow-direct-references = true


[project.optional-dependencies]
rocm = [
    "torch==2.9.1+rocm7.2.0.lw.git7e1940d4",
    "torchaudio==2.9.0+rocm7.2.0.gite3c6ee2b",
    "triton==3.5.1+rocm7.2.0.gita272dfa8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.24",
    "types-PyYAML>=6.0.12.20250915",
    "typing-extensions>=4.15.0",
    "docstr-coverage>=2.3.2",
    "interrogate>=1.7.0",
]
bench = [
    "pyamdgpuinfo==2.1.7",
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
torch = [
    { index = "rocm-wheels-7-2", extra = "rocm" },
]
torchaudio = [
    { index = "rocm-wheels-7-2", extra = "rocm" },
]
triton = [
    { index = "rocm-wheels-7-2", extra = "rocm" },
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "rocm-wheels-7-2"
url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.2/"
format = "flat"
explicit = true

# ------------------------------
#   Ruff (linter + formatter)
# ------------------------------
[tool.ruff]
line-length = 88
target-version = "py310"                # adjust if you target 3.9/3.12
preview = true                          # needed for DOC + FA rules
src = ["insanely_fast_whisper_rocm"]     # first-party package path


[tool.ruff.lint]
select = [
  "F",      # Pyflakes (includes F403/F405 for wildcard imports)
  "E", "W", # pycodestyle subset (PEP 8 basics)
  "N",      # pep8-naming (snake_case, PascalCase)
  "I",      # isort (import sorting)
  "D",      # pydocstyle (docstring presence + Google style)
  "DOC",    # pydoclint (Args/Returns/Raises checks; needs preview)
  "ANN",    # flake8-annotations (type hints everywhere)
  "TID",    # flake8-tidy-imports (absolute imports: TID252)
  "UP",     # pyupgrade (modern syntax)
  "FA",     # future-annotations (auto `from __future__`; needs preview)
]

# Optional: stricter typing rules (discourage bare Any)
extend-select = ["ANN401"]


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruff.lint.isort]
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder",
]
known-first-party = ["insanely_fast_whisper_rocm"]


[tool.ruff.lint.per-file-ignores]
# Temporarily ignore long-line checks in tests
"tests/**/*.py" = ["E501"]

# ------------------------------
#       Pytest & Coverage
# ------------------------------
[tool.pytest.ini_options]
# addopts = "-q --cov=insanely_fast_whisper_rocm --cov-branch --cov-report=term-missing:skip-covered --cov-report=xml --cov-report=html --cov-fail-under=85"
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore:websockets\\.legacy is deprecated;.*:DeprecationWarning",
]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: mark tests that hit real external integrations or slower paths",
    "gpu: mark tests that require a working ROCm GPU runtime",
    "real_media: mark tests that require an opt-in local real media file",
    "webui: mark tests that exercise the Gradio WebUI",
]


[tool.coverage.run]
omit = [
    "*/__main__.py",
]

[tool.ty.environment]
python-version = "3.10"
python-platform = "linux"

[tool.ty.src]
include = ["insanely_fast_whisper_rocm"]
exclude = [
    "**/__pycache__",
    "**/.venv",
    "**/venv",
    "**/node_modules",
    "**/temp_uploads",
    "**/transcripts",
    "**/transcripts_zips",
]

[tool.ty.terminal]
output-format = "concise"
